backend/sample-agent/pyproject.toml
[project]
name = "sample-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.107.2",
    "python-dotenv>=1.1.1",
]
